feat: Add deprecated hydro_metrics for backward compatibility
- Introduced a new module `hydro_metrics.py` to provide backward compatibility for legacy hydrological metrics.
- Added a modern wrapper function `calculate_hydrological_indices()` to calculate comprehensive hydrological indices.
- Deprecated the old `hydro_job()` function, recommending users to switch to the new wrapper.
- Included legacy metric functions (`nse`, `kge`, `rmse`) with warnings to guide users to modern alternatives.
- Documented the changes and provided examples for better usability.

+++ src/hydro/base_flow.py
@@ -122,8 +122,9 @@ def _reflect_series(discharge: np.ndarray, reflect_points: int) -> np.ndarray:
     Returns:
         Reflected discharge array
     """
+    # Always return float64 to ensure consistent type
     if len(discharge) - 1 <= reflect_points:
-        return discharge.copy()
+        return discharge.astype(np.float64).copy()
 
     q_reflect = np.full(len(discharge) + 2 * reflect_points, np.nan, dtype=np.float64)
